requirements.txt
python-binance>=1.0.19
click>=8.1.0
python-dotenv>=1.0.0
